Reworked returns from function to return object from class ReturnBody and reworked javascripts to append new data, not to reload them

+++ public_html/app/controller/AccountController.php
@@ -18,17 +18,16 @@ class AccountController
             $account = new Account($_POST['name'], $_POST['current_amount'], $_SESSION['logged_user']);
             $accountDAO = new AccountDAO();
             if (!Validator::validateName($account->getName())) {
-                throw new BadRequestException("Name must be have greater than " . MIN_LENGTH_NAME . " symbols");
+                throw new BadRequestException("Name must be have greater than " . MIN_LENGTH_NAME . " symbols!");
             } elseif (!Validator::validateAmount($account->getCurrentAmount())) {
-                throw new BadRequestException("Amount must be between 0 and " . MAX_AMOUNT . " inclusive");
+                throw new BadRequestException("Amount must be between 0 and " . MAX_AMOUNT . " inclusive!");
             } else {
                 $id = $accountDAO->createAccount($account);
                 $account->setId($id);
+                return new ResponseBody("Account created successfully!", $account);
             }
-        } else {
-            throw new BadRequestException("Bad request");
         }
-        return new ResponseBody("Account created successfully", $account);
+        throw new BadRequestException("Bad request.");
     }
 
     public function getAll() {
@@ -37,12 +36,13 @@ class AccountController
             $accounts = $accountsDAO->getMyAccounts($_SESSION['logged_user']);
             return new ResponseBody(null, $accounts);
         }
+        throw new BadRequestException("Bad request.");
     }
 
     public function edit() {
         if (isset($_POST["edit"])) {
             if (!Validator::validateName($_POST["name"])) {
-                throw new BadRequestException("Name must be have between " . MIN_LENGTH_NAME . " and ". " symbols");
+                throw new BadRequestException("Name must be have between " . MIN_LENGTH_NAME . " and ". " symbols!");
             }
             $account_id = $_POST["account_id"];
             $accountDAO = new AccountDAO();
@@ -50,16 +50,18 @@ class AccountController
             //** @Account $account */
             $account = $accountDAO->getAccountById($account_id);
             if (!$account) {
-                throw new ForbiddenException("This account is not yours");
+                throw new ForbiddenException("This account is not yours!");
             }
-            $newAccountName = new Account($_POST["name"], $account->getCurrentAmount(), $_SESSION['logged_user']);
-            $newAccountName->setId($account->getId());
-            if ($newAccountName->getOwnerId() == $account->getOwnerId()) {
-                $accountDAO->editAccount($newAccountName);
+            $account->setName($_POST['name']);
+
+            if ($account->getOwnerId() == $_SESSION['logged_user']) {
+                $accountDAO->editAccount($account);
+                return new ResponseBody('Account edited successfully!', $account);
             } else {
-                throw new ForbiddenException("This account is not yours");
+                throw new ForbiddenException("This account is not yours!");
             }
         }
+        throw new BadRequestException("Bad request.");
     }
 
     public function delete() {
@@ -68,11 +70,13 @@ class AccountController
             $accountDAO = new AccountDAO();
             $account = $accountDAO->getAccountById($account_id);
 
-            if ($account->getOwnerId() == $_SESSION['logged_user']) {
+            if ($account && $account->getOwnerId() == $_SESSION['logged_user']) {
                 $accountDAO->deleteAccount($account->getId());
+                return new ResponseBody("Account delete successfully!", $account);
             } else {
-                throw new ForbiddenException("This account is not yours");
+                throw new ForbiddenException("This account is not yours!");
             }
         }
+        throw new BadRequestException("Bad request.");
     }
 }
